new site ^^

+++ mariadb/db_init/omega.sql
@@ -1,54 +1,54 @@
 
 CREATE TABLE IF NOT EXISTS coa (
-	campus_id	INTEGER NOT NULL,
-	intra_id	INTEGER NOT NULL,
-	guild_id	INTEGER NOT NULL,
-	discord_id	INTEGER NOT NULL
+	campus_id	BIGINT NOT NULL,
+	intra_id	BIGINT NOT NULL,
+	guild_id	BIGINT NOT NULL,
+	discord_id	BIGINT NOT NULL
 );
 CREATE TABLE IF NOT EXISTS cursus (
-	campus_id	INTEGER NOT NULL,
-	intra_id	INTEGER NOT NULL,
-	guild_id	INTEGER NOT NULL,
-	discord_id	INTEGER NOT NULL
+	campus_id	BIGINT NOT NULL,
+	intra_id	BIGINT NOT NULL,
+	guild_id	BIGINT NOT NULL,
+	discord_id	BIGINT NOT NULL
 );
 CREATE TABLE IF NOT EXISTS groups (
-	campus_id	INTEGER NOT NULL,
-	intra_id	INTEGER NOT NULL,
-	guild_id	INTEGER NOT NULL,
-	discord_id	INTEGER NOT NULL
+	campus_id	BIGINT NOT NULL,
+	intra_id	BIGINT NOT NULL,
+	guild_id	BIGINT NOT NULL,
+	discord_id	BIGINT NOT NULL
 );
 CREATE TABLE IF NOT EXISTS users (
-	omega_id	INTEGER NOT NULL AUTO_INCREMENT,
-	discord_id	INTEGER NOT NULL,
+	omega_id	BIGINT NOT NULL AUTO_INCREMENT,
+	discord_id	BIGINT NOT NULL,
 	intra_id	TEXT NOT NULL,
 	PRIMARY KEY(omega_id)
 );
 CREATE TABLE IF NOT EXISTS years (
-	campus_id	INTEGER NOT NULL,
-	intra_id	INTEGER NOT NULL,
-	guild_id	INTEGER NOT NULL,
-	discord_id	INTEGER NOT NULL
+	campus_id	BIGINT NOT NULL,
+	intra_id	BIGINT NOT NULL,
+	guild_id	BIGINT NOT NULL,
+	discord_id	BIGINT NOT NULL
 );
 CREATE TABLE IF NOT EXISTS nick (
-	campus_id	INTEGER NOT NULL,
-	guild_id	INTEGER NOT NULL,
+	campus_id	BIGINT NOT NULL,
+	guild_id	BIGINT NOT NULL,
 	format	TEXT NOT NULL
 );
 CREATE TABLE IF NOT EXISTS project (
-	campus_id	INTEGER NOT NULL,
-	intra_id	INTEGER NOT NULL,
-	in_progress	INTEGER NOT NULL,
-	finished	INTEGER NOT NULL,
-	validated	INTEGER NOT NULL,
-	discord_id	INTEGER NOT NULL,
-	guild_id	INTEGER NOT NULL
+	campus_id	BIGINT NOT NULL,
+	intra_id	BIGINT NOT NULL,
+	in_progress	BIGINT NOT NULL,
+	finished	BIGINT NOT NULL,
+	validated	BIGINT NOT NULL,
+	discord_id	BIGINT NOT NULL,
+	guild_id	BIGINT NOT NULL
 );
 CREATE TABLE IF NOT EXISTS temp_auth (
-	discord_id	INTEGER NOT NULL,
+	discord_id	BIGINT NOT NULL,
 	code	TEXT NOT NULL
 );
 CREATE TABLE IF NOT EXISTS new_users (
-	discord_id	INTEGER NOT NULL,
+	discord_id	BIGINT NOT NULL,
 	intra_id	TEXT NOT NULL
 );
 CREATE TABLE IF NOT EXISTS maintenance (
@@ -56,8 +56,8 @@ CREATE TABLE IF NOT EXISTS maintenance (
 	status	TEXT NOT NULL
 );
 CREATE TABLE IF NOT EXISTS ticket (
-	user_id	INTEGER NOT NULL,
-	channel_id	INTEGER NOT NULL
+	user_id	BIGINT NOT NULL,
+	channel_id	BIGINT NOT NULL
 );
 CREATE TABLE IF NOT EXISTS status (
 	name	TEXT NOT NULL
